PAL of do_exec_* APIs
