+++ contrib/win32/openssh/ssh.vcxproj.filters
@@ -285,6 +285,9 @@
     <ClInclude Include="$(OpenSSH-Src-Path)xmalloc.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="$(OpenSSH-Src-Path)sshdoexec.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="$(OpenSSH-Src-Path)clientloop.c">
